fix(ble): address fresh-context review findings
- phone emulator: decode_rsc returned a 2-tuple on the short-packet path
  but the caller unpacks 3 values -> return None, None, None (crash fix)
- driver: drop dead `resetCb` member (wired callback is the file-static
  s_resetCb); clears a misleading DRY wrinkle
- rsc_codec: accurate speed=0.0 m/s comment; clamp cadence field to 255
- driver: use uint64_t for cadence math to avoid theoretical overflow
- main: drive LED off when BLE.begin() fails so it is distinct from advertising

All changes keep pio test -e native (34/34) and pio run -e nanorp2040connect
green.

+++ src/ble_peripheral_arduino.cpp
@@ -101,7 +101,9 @@ void ArduinoBlePeripheral::notifySteps(uint32_t totalSteps) {
                              : 0;
         unsigned long dtSec = (now - lastNotifyMs) / 1000;
         if (dtSec > 0) {
-            cadenceSpm = static_cast<uint16_t>(delta * 60 / dtSec);
+            // uint64_t to avoid overflow when delta is large (defensive).
+            cadenceSpm = static_cast<uint16_t>(
+                static_cast<uint64_t>(delta) * 60 / dtSec);
         }
     }
     lastTotalSteps = totalSteps;
